platform: move GenerateFakeKey to platform/util
GenerateFakeKey was added for DigitalOcean but it is now needed by Azure
as well. Move it to a centralized area and update Digitalocean to use
the new location.
